requirements/base.txt
# Base
pytz==2020.1
python-slugify==4.0.0
Pillow==7.1.2
psycopg2==2.8.5 --no-binary psycopg2

# Django
django==3.0.7

# Django REST Framework
djangorestframework==3.11.0

# Django filter
django-filter==2.3.0

# Django REST Framework Simple Json Web Token
djangorestframework-simplejwt==4.4.0

# pyJWT
pyjwt==1.7.1

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==5.1.0

# Celery
redis==3.5.3
django-redis==4.12.1
celery==4.4.6
flower==0.9.4
